chore: add policy api unit tests
