--- api/policies.kyverno.io/v1alpha1/validating_spec_types.go
@@ -1,172 +1,0 @@
-package v1alpha1
-
-import (
-	admissionregistrationv1 "k8s.io/api/admissionregistration/v1"
-)
-
-// ValidatingPolicySpec is the specification of the desired behavior of the ValidatingPolicy.
-type ValidatingPolicySpec struct {
-	// MatchConstraints specifies what resources this policy is designed to validate.
-	// The AdmissionPolicy cares about a request if it matches _all_ Constraints.
-	// However, in order to prevent clusters from being put into an unstable state that cannot be recovered from via the API
-	// ValidatingAdmissionPolicy cannot match ValidatingAdmissionPolicy and ValidatingAdmissionPolicyBinding.
-	// Required.
-	MatchConstraints *admissionregistrationv1.MatchResources `json:"matchConstraints,omitempty"`
-
-	// Validations contain CEL expressions which is used to apply the validation.
-	// Validations and AuditAnnotations may not both be empty; a minimum of one Validations or AuditAnnotations is
-	// required.
-	// +listType=atomic
-	// +optional
-	Validations []admissionregistrationv1.Validation `json:"validations,omitempty"`
-
-	// failurePolicy defines how to handle failures for the admission policy. Failures can
-	// occur from CEL expression parse errors, type check errors, runtime errors and invalid
-	// or mis-configured policy definitions or bindings.
-	//
-	// A policy is invalid if spec.paramKind refers to a non-existent Kind.
-	// A binding is invalid if spec.paramRef.name refers to a non-existent resource.
-	//
-	// failurePolicy does not define how validations that evaluate to false are handled.
-	//
-	// When failurePolicy is set to Fail, ValidatingAdmissionPolicyBinding validationActions
-	// define how failures are enforced.
-	//
-	// Allowed values are Ignore or Fail. Defaults to Fail.
-	// +optional
-	FailurePolicy *admissionregistrationv1.FailurePolicyType `json:"failurePolicy,omitempty"`
-
-	// auditAnnotations contains CEL expressions which are used to produce audit
-	// annotations for the audit event of the API request.
-	// validations and auditAnnotations may not both be empty; a least one of validations or auditAnnotations is
-	// required.
-	// +listType=atomic
-	// +optional
-	AuditAnnotations []admissionregistrationv1.AuditAnnotation `json:"auditAnnotations,omitempty"`
-
-	// MatchConditions is a list of conditions that must be met for a request to be validated.
-	// Match conditions filter requests that have already been matched by the rules,
-	// namespaceSelector, and objectSelector. An empty list of matchConditions matches all requests.
-	// There are a maximum of 64 match conditions allowed.
-	//
-	// If a parameter object is provided, it can be accessed via the `params` handle in the same
-	// manner as validation expressions.
-	//
-	// The exact matching logic is (in order):
-	//   1. If ANY matchCondition evaluates to FALSE, the policy is skipped.
-	//   2. If ALL matchConditions evaluate to TRUE, the policy is evaluated.
-	//   3. If any matchCondition evaluates to an error (but none are FALSE):
-	//      - If failurePolicy=Fail, reject the request
-	//      - If failurePolicy=Ignore, the policy is skipped
-	//
-	// +patchMergeKey=name
-	// +patchStrategy=merge
-	// +listType=map
-	// +listMapKey=name
-	// +optional
-	MatchConditions []admissionregistrationv1.MatchCondition `json:"matchConditions,omitempty" patchStrategy:"merge" patchMergeKey:"name"`
-
-	// Variables contain definitions of variables that can be used in composition of other expressions.
-	// Each variable is defined as a named CEL expression.
-	// The variables defined here will be available under `variables` in other expressions of the policy
-	// except MatchConditions because MatchConditions are evaluated before the rest of the policy.
-	//
-	// The expression of a variable can refer to other variables defined earlier in the list but not those after.
-	// Thus, Variables must be sorted by the order of first appearance and acyclic.
-	// +patchMergeKey=name
-	// +patchStrategy=merge
-	// +listType=map
-	// +listMapKey=name
-	// +optional
-	Variables []admissionregistrationv1.Variable `json:"variables,omitempty" patchStrategy:"merge" patchMergeKey:"name"`
-
-	// Generate specifies whether to generate a Kubernetes ValidatingAdmissionPolicy.
-	// Optional. Defaults to "false" if not specified.
-	// +optional
-	// +kubebuilder:default=false
-	Generate *bool `json:"generate,omitempty"`
-
-	// ValidationAction specifies the action to be taken when the matched resource violates the policy.
-	// Required.
-	// +listType=set
-	ValidationAction []admissionregistrationv1.ValidationAction `json:"validationActions,omitempty"`
-
-	// WebhookConfiguration defines the configuration for the webhook.
-	// +optional
-	WebhookConfiguration *WebhookConfiguration `json:"webhookConfiguration,omitempty"`
-
-	// EvaluationConfiguration defines the configuration for the policy evaluation.
-	// +optional
-	EvaluationConfiguration *EvaluationConfiguration `json:"evaluation,omitempty"`
-}
-
-// AdmissionEnabled checks if admission is set to true
-func (s ValidatingPolicySpec) AdmissionEnabled() bool {
-	if s.EvaluationConfiguration == nil || s.EvaluationConfiguration.Admission == nil || s.EvaluationConfiguration.Admission.Enabled == nil {
-		return true
-	}
-	return *s.EvaluationConfiguration.Admission.Enabled
-}
-
-// BackgroundEnabled checks if background is set to true
-func (s ValidatingPolicySpec) BackgroundEnabled() bool {
-	if s.EvaluationConfiguration == nil || s.EvaluationConfiguration.Background == nil || s.EvaluationConfiguration.Background.Enabled == nil {
-		return true
-	}
-	return *s.EvaluationConfiguration.Background.Enabled
-}
-
-// EvaluationMode returns the evaluation mode of the policy.
-func (s ValidatingPolicySpec) EvaluationMode() EvaluationMode {
-	if s.EvaluationConfiguration == nil || s.EvaluationConfiguration.Mode == "" {
-		return EvaluationModeKubernetes
-	}
-	return s.EvaluationConfiguration.Mode
-}
-
-type WebhookConfiguration struct {
-	// TimeoutSeconds specifies the maximum time in seconds allowed to apply this policy.
-	// After the configured time expires, the admission request may fail, or may simply ignore the policy results,
-	// based on the failure policy. The default timeout is 10s, the value must be between 1 and 30 seconds.
-	TimeoutSeconds *int32 `json:"timeoutSeconds,omitempty"`
-}
-
-type EvaluationConfiguration struct {
-	// Mode is the mode of policy evaluation.
-	// Allowed values are "Kubernetes" or "JSON".
-	// Optional. Default value is "Kubernetes".
-	// +optional
-	Mode EvaluationMode `json:"mode,omitempty"`
-
-	// Admission controls policy evaluation during admission.
-	// +optional
-	Admission *AdmissionConfiguration `json:"admission,omitempty"`
-
-	// Background  controls policy evaluation during background scan.
-	// +optional
-	Background *BackgroundConfiguration `json:"background,omitempty"`
-}
-
-type AdmissionConfiguration struct {
-	// Enabled controls if rules are applied during admission.
-	// Optional. Default value is "true".
-	// +optional
-	// +kubebuilder:default=true
-	Enabled *bool `json:"enabled,omitempty"`
-}
-
-type BackgroundConfiguration struct {
-	// Enabled controls if rules are applied to existing resources during a background scan.
-	// Optional. Default value is "true". The value must be set to "false" if the policy rule
-	// uses variables that are only available in the admission review request (e.g. user name).
-	// +optional
-	// +kubebuilder:default=true
-	Enabled *bool `json:"enabled,omitempty"`
-}
-
-type EvaluationMode string
-
-const (
-	EvaluationModeKubernetes EvaluationMode = "Kubernetes"
-	EvaluationModeJSON       EvaluationMode = "JSON"
-)
